requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
pydantic==2.7.4
pydantic-settings==2.3.4
sqlalchemy==2.0.30
pillow==10.3.0
httpx==0.27.0
pytest==8.3.2
